Support overriding the table alias name
Continue https://github.com/clio/jit_preloader/pull/18

When the association scope is chained with conditions of other tables, the names of join tables are also added to `references_values`. It becomes impossible to identify the table alias name from the `references_values`.

The PR provides a way to customize the table alias name when the default way cannot figure out the correct value.